Split test-utils tests into a folder and drop its max-lines exemption
The test-utils self-test file had grown to 1,773 lines and was
grandfathered into biome's noExcessiveLinesPerFile exemption list. Split
it into test/lib/test-utils/ — grouped by area (stubs/caches/request mocks,
csrf & form helpers, ticket & join flows, listing/attendee factories, error
paths & contracts, and db-backed & settings contracts) — so every file
sits under 400 lines, then removed the file's exemption entry.

The wrapper-scoped expectFormPostWithBody helper travels with the one shard
that uses it. Bringing the file back under jscpd's per-file scan cap
surfaced two duplicated ticket-form blocks, now the local helpers
expectSingleAttendeeQuantity and payableListing. All original tests carry
over unchanged and pass (207 across the folder); jscpd is back to 0%.
